feat: seed store
