+++ lib/src/wallet/stores/in_memory_key_store.dart
@@ -1,11 +1,8 @@
-import 'dart:typed_data';
-
 import 'key_store_interface.dart';
 import 'stored_key.dart';
 
 class InMemoryKeyStore implements KeyStore {
   final Map<String, StoredKey> _keyPairStore = {};
-  Uint8List? _seed;
 
   @override
   Future<void> set(String key, StoredKey value) async {
@@ -15,16 +12,6 @@ class InMemoryKeyStore implements KeyStore {
   @override
   Future<StoredKey?> get(String key) async {
     return _keyPairStore[key];
-  }
-
-  @override
-  Future<void> setSeed(Uint8List seed) async {
-    _seed = seed;
-  }
-
-  @override
-  Future<Uint8List?> getSeed() async {
-    return _seed;
   }
 
   @override
@@ -40,6 +27,5 @@ class InMemoryKeyStore implements KeyStore {
   @override
   Future<void> clear() async {
     _keyPairStore.clear();
-    _seed = null;
   }
 }
