+++ lib/src/wallet/stores/key_store_interface.dart
@@ -1,5 +1,3 @@
-import 'dart:typed_data';
-
 import 'stored_key.dart';
 
 /// An interface for a key-value storage.
@@ -10,13 +8,6 @@ abstract class KeyStore {
   /// Retrieves the value associated with the given key.
   /// Returns null if the key does not exist or stores a seed.
   Future<StoredKey?> get(String key);
-
-  /// Stores the seed value, overwriting any previous seed.
-  Future<void> setSeed(Uint8List seed);
-
-  /// Retrieves the stored seed value.
-  /// Returns null if no seed has been stored.
-  Future<Uint8List?> getSeed();
 
   /// Removes the key pair associated with the given key. Does not affect the seed.
   Future<void> remove(String key);
