+++ test/wallet/persistent_wallet_test.dart
@@ -1,6 +1,6 @@
 import 'dart:typed_data';
 
-import 'package:ssi/src/wallet/key_store/in_memory_key_store.dart';
+import 'package:ssi/src/wallet/stores/in_memory_key_store.dart';
 import 'package:ssi/ssi.dart';
 import 'package:test/test.dart';
